requirements.txt
aniso8601==9.0.1
boto3==1.17.87
botocore==1.20.87
catboost==0.26
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
graphviz==0.16
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-simplexml==0.1.5
pytz==2021.1
regex==2021.4.4
retrying==1.3.3
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.61.0
urllib3==1.26.5
Werkzeug==2.0.1
